Rename and fix country_spec.rb
* Was country.rb, wasn't running
* Would have failed if it was
